% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsft.input.R
\name{as.data.frame.rpsft.input}
\alias{as.data.frame.rpsft.input}
\title{as.data.frame.rpsft.input}
\usage{
\method{as.data.frame}{rpsft.input}(x)
}
\arguments{
\item{x}{An rpsft.input object created with rpsft.input()}
}
\description{
Converts an rpsft.input object to a data frame.
}
\examples{

sim.df <- simStudy()
x <- rpsft.input(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
as.data.frame(x)
}
\keyword{RPSFT,}
\keyword{Survival}
