% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvCF.R
\name{as.data.frame.SurvCF}
\alias{as.data.frame.SurvCF}
\title{as.data.frame.SurvCF}
\usage{
\method{as.data.frame}{SurvCF}(x)
}
\arguments{
\item{x}{An SurvCF object created with SurvCF()}
}
\description{
Converts an SurvCF object to a data frame.
}
\examples{

sim.df <- simStudy()
x <- SurvExt(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
as.data.frame(x)
}
\keyword{RPSFT,}
\keyword{Survival}
