% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvExt.R
\name{as.data.frame.SurvExt}
\alias{as.data.frame.SurvExt}
\title{as.data.frame.SurvExt}
\usage{
\method{as.data.frame}{SurvExt}(x)
}
\arguments{
\item{x}{An SurvExt object created with SurvExt()}
}
\description{
Converts an SurvExt object to a data frame.
}
\examples{

sim.df <- simStudy()
x <- SurvExt(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
as.data.frame(x)
}
\keyword{RPSFT,}
\keyword{Survival}
