% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.RPSFT.R
\name{cox.RPSFT}
\alias{cox.RPSFT}
\title{cox.RPSFT}
\usage{
cox.RPSFT(x, use.latent.only = FALSE)
}
\arguments{
\item{x}{An RPSFT object created with RPSFT()}

\item{use.latent.only}{If TRUE experimental arm is imputed from latent time. If FALSE then observed times used for experimental arm.}
}
\description{
Annalyzes a RPSFT object using a cox model. Corrects the standard errors based on rank test statistic used in G estimation.
}
\examples{

sim.df <- simStudy()
x <- rpsft.input(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
y <- RPSFT(x)
cox.RPSFT(y)
}
\keyword{RPSFT,}
\keyword{Survival}
