% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsft.input.R
\name{plot.rpsft.input}
\alias{plot.rpsft.input}
\title{plot.rpsft.input}
\usage{
\method{plot}{rpsft.input}(x)
}
\arguments{
\item{x}{An rpsft.input object created with rpsft.input()}
}
\description{
Plots an rpsft.input object using GGally and some default settings.
}
\examples{

sim.df <- simStudy()
x <- rpsft.input(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
plot(x)
}
\keyword{RPSFT,}
\keyword{Survival}
