% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsft.input.R
\name{plot.SurvExt}
\alias{plot.SurvExt}
\title{plot.SurvExt}
\usage{
\method{plot}{SurvExt}(x)
}
\arguments{
\item{x}{An SurvExt object created with SurvExt()}
}
\description{
Plots a SurvExt object using GGally and some default settings.
}
\examples{

sim.df <- simStudy()
x <- SurvExt(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
plot(x)
}
\keyword{RPSFT,}
\keyword{Survival}
