% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPSFT.R
\name{print.RPSFT}
\alias{print.RPSFT}
\title{print.RPSFT}
\usage{
\method{print}{RPSFT}(x)
}
\arguments{
\item{x}{An RPSFT object created with RPSFT()}
}
\description{
Prints a summary of an RPSFT object to screen.
}
\examples{

sim.df <- simStudy()
x <- rpsft.input(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
y <- RPSFT(x)
y
}
\keyword{RPSFT,}
\keyword{Survival}
