% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvExt.R
\name{print.SurvExt}
\alias{print.SurvExt}
\title{print.SurvExt}
\usage{
\method{print}{SurvExt}(x)
}
\arguments{
\item{x}{An rpsft.input object created with rpsft.input()}
}
\description{
Prints a summary of n rpsft.input object to screen.
}
\examples{

sim.df <- simStudy()
x <- SurvExt(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
x
}
\keyword{RPSFT,}
\keyword{Survival}
