% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsft.input.R
\name{rpsft.input}
\alias{rpsft.input}
\title{rpsft.input}
\usage{
rpsft.input(formula, data, Exposure, AdminCensTime)
}
\arguments{
\item{formula}{A formula for survival time on randomized treatment. Can only include single regression variable.}

\item{data}{Required. A data frame in which to interpret the variables named in the formula and other arguments.}

\item{Exposure}{Defines the exposure to study drug per patient.}

\item{AdminCensTime}{Defined the administrative censoring time.}
}
\description{
Creates an rpsft.input object that combines surival and exposure information.
}
\examples{

sim.df <- simStudy()
x <- rpsft.input(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
plot(x)
RPSFT(x)
}
\keyword{RPSFT,}
\keyword{Survival}
