% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPSFT.R
\name{RPSFT}
\alias{RPSFT}
\title{A function to do RPSFT}
\usage{
RPSFT(rpsft.input, psimat = matrix(ncol = 1, data = seq(from = -3, to = 3, by
  = 0.1)), Grho = 0, twopass = TRUE)
}
\arguments{
\item{rpsft.input}{An object created using rpsft.input()}

\item{psimat}{A matrix of values to try in grid search}

\item{Grho}{Defines the test used in the grid search (0 = logrank, 1 = wilcoxon)}

\item{twopass}{Defines if two searches are performed (wide then fine)}
}
\description{
This function applies RPSFT method
}
\examples{

sim.df <- simStudy()
x <- rpsft.input(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
plot(x)
RPSFT(x)
}
\keyword{RPSFT,}
\keyword{Survival}
