% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy.R
\name{simStudy}
\alias{simStudy}
\title{A function to simulate Survival data}
\usage{
simStudy(simid = 1, seed = 1234, rho = 0, p.switch = 0.25,
  prop.pd.int = 0, beta_1a = log(0.7), beta_1b = log(0.7),
  beta_2a = log(0.7), beta_2b = log(0.7), beta_pd = log(0.4),
  arm.n = 250, enroll.start = 0, enroll.end = 2, admin.censor = 3,
  os.gamma = 1.2, os.lambda = 0.3, ttp.gamma = 1.5, ttp.lambda = 2,
  include.metadata = FALSE)
}
\arguments{
\item{simid}{Identifer added to simulated dataframe. Defaults to 1.}

\item{seed}{Seed used for random number generator. Defaults to 1234.}

\item{rho}{correlation coefficient between TTP and OS. Defaults to 0.}

\item{p.switch}{proportion of patients who switch. Defaults to 0.25.}

\item{prop.pd.int}{proportion of patients with PFS before switch allowed. Defaults to 0.}

\item{beta_1a}{treatment effect (as log(Hazard Ratio)). Defaults to log(0.7).}

\item{beta_1b}{treatment effect (as log(Hazard Ratio)). Defaults to log(0.7).}

\item{beta_2a}{treatment effect (as log(Hazard Ratio)). Defaults to log(0.7).}

\item{beta_2b}{treatment effect (as log(Hazard Ratio)). Defaults to log(0.7).}

\item{beta_pd}{treatment effect on prog (as log(HR)). Defaults to log(0.4).}

\item{arm.n}{patients per arm. Defaults to 250.}

\item{enroll.start}{start of enrollment. Defaults to 0.}

\item{enroll.end}{end of enrollment. Defaults to 2.}

\item{admin.censor}{end of trial. Defaults to 3.}

\item{os.gamma}{weibull shape - for OS. Defaults to 1.2.}

\item{os.lambda}{weibull scale - for OS. Defaults to 0.3.}

\item{ttp.gamma}{weibull shape - for TTP. Defaults to 1.5.}

\item{ttp.lambda}{weibull scale - for TTP. Defaults to 2.}

\item{include.metadata}{Should returned dataframe include simulation parameters. Defaults to False.}
}
\description{
This function simulates survival data with correlated time to progression and overall survival times.
}
\examples{
require(survival)
require(dplyr)

tst.df <- simStudy()
survfit(Surv(pfs.t, event = pfs.e) ~ x.trt, data = tst.df) \%>\%
  plot()

survfit(Surv(os.t, event = os.e) ~ x.trt, data = tst.df) \%>\%
  plot()
}
\keyword{RPSFT,}
\keyword{Simulation}
\keyword{Survival,}
