% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvCF.R
\name{SurvCF}
\alias{SurvCF}
\title{SurvCF - Counter Factual Survival Object}
\usage{
SurvCF(x, CFExposure, UseObserved = FALSE)
}
\arguments{
\item{x}{An RPSFT object created with RPSFT().}

\item{CFExposure}{Defines the counter factual exposure to study drug per patient.}

\item{UseObserved}{TRUE/FALSE indicator per patient if Observed values should be used for a patient.}
}
\description{
Creates an SurvCF object that combines survival and counter factual exposure information. Derived from an RPSFT model.
}
\examples{

sim.df <- simStudy()
x <- SurvExt(Surv(os.t, os.e) ~ I(x.trt==1),
   CFExposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
y <- RPSFT(x)
z <- SurvCF(y, CFExposure = ifelse(sim.df$x.trt == 0, 0, sim.df$os.t))
z <- SurvCF(y, CFExposure = 0, UseObserved = sim.df$x.trt == 1)
}
\keyword{RPSFT,}
\keyword{Survival}
