% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvExt.R
\name{SurvExt}
\alias{SurvExt}
\title{SurvExt - Extended Survival Object}
\usage{
SurvExt(formula, data, Exposure, AdminCensTime, FirstSwitchTime)
}
\arguments{
\item{formula}{A formula for survival time on randomized treatment. Can only include single regression variable.}

\item{data}{Required. A data frame in which to interpret the variables named in the formula and other arguments.}

\item{Exposure}{Defines the exposure to study drug per patient.}

\item{AdminCensTime}{Defines the administrative censoring time.}

\item{FirstSwitchTime}{Defines the first time of switch}
}
\description{
Creates an SurvExt object that combines survival and exposure information. Can be used for RPSFT analysis.
}
\examples{

sim.df <- simStudy()
x <- SurvExt(Surv(os.t, os.e) ~ I(x.trt==1),
   Exposure = ifelse(x.trt == 1, os.t, ifelse(x.switch == 1, os.t - t.switch, 0)),
   AdminCensTime = t.censor,
   data = sim.df)
plot(x)
RPSFT(x)
}
\keyword{RPSFT,}
\keyword{Survival}
